Begin migrating towards timestamp-based activity logs
+ query database for timestamp-mapped card activity
+ work around cal-heatmap UTC limitations

Up to this point we were following Anki's example in working with
activity mapped to days relative to a specific date (in case of
history: collection creation date, in case of forecast: today).
We would only switch to absolute timestamps in preparation of passing
the data on to cal-heatmap.

While this approach works well for Anki's own statistics module, in the
case of Review Heatmap it introduced a lot of added complexity that
contributed to date mismatches and other time-related bugs over
the years. Even now, Review Heatmap still maps user activity to
incorrect dates under some circumstances. Because of both timezones,
DST properties, and Anki's own daily cutoff setting factoring into
these situations, bugs like this are notoriously hard to track down.

With this commit we begin migrating towards a different approach which
should be more resilient to date- and time-related bugs. The goal is
to eliminate these issues once and for all.