Removed inverted colors in night mode
Night mode is still used, eg for keeping empty tiles black instead of white.
But for color schemes, a 'small' number is assigned a light color while a
'large' number is assigned a darker color. There's no default inversion for
night mode. I've left the CSS in, so maybe we can add a checkbox in Options
to let the user choose Inverted colors if they'd like.